moved DocketTypeBadge file to components
